combine timeline date and timeline component [wip]
